make service class from requests.js
